Backend1/requirements.txt
chromadb==0.5.5
Flask==3.0.3
flask_cors==5.0.0
langchain==0.2.16
langchain_community==0.2.16
langchain_core==0.2.38
langchain_groq==0.1.9
Markdown==3.7
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
SQLAlchemy==2.0.30
Werkzeug==3.0.4
pymysql==1.1.1